Added validation methods

+++ flask-lessons/trello-clone/blueprints/cards_bp.py
@@ -2,14 +2,15 @@ from flask import Blueprint, request, abort
 from setup import db
 from models.card import Card, CardSchema
 from flask_jwt_extended import jwt_required, get_jwt_identity
-from auth import admin_required
+from auth import authorize
+from blueprints.comment_bp import comments_bp
 
 cards_bp = Blueprint('cards', __name__, url_prefix='/cards')
 
 @cards_bp.route('/')
 @jwt_required()
 def all_cards():
-    admin_required()
+    authorize()
     # select * from cards;
     stmt = db.select(
         Card
@@ -21,7 +22,7 @@ def all_cards():
 @cards_bp.route('/<int:id>/')
 @jwt_required()
 def one_card(id):
-    admin_required()
+    authorize()
     stmt = db.select(Card).filter_by(id=id)
     card = (db.session.scalar(stmt))
     if card: # If card is a truthy value, return
@@ -48,11 +49,11 @@ def create_card():
 @cards_bp.route('/<int:id>', methods=['PUT','PATCH'])
 @jwt_required()
 def update_card(id):
-    admin_required()
     card_info = CardSchema(exclude=['id','date_created']).load(request.json)
     stmt = db.select(Card).filter_by(id=id)
     card = db.session.scalar(stmt)
     if card: # If card is a truthy value, return
+        authorize(card.user_id)
         card.title = card_info.get('title', card.title)
         card.description = card_info.get('description', card.description)
         card.status = card_info.get('status', card.status)
@@ -65,12 +66,14 @@ def update_card(id):
 @cards_bp.route('/<int:id>', methods=['DELETE'])
 @jwt_required()
 def delete_card(id):
-    admin_required()
     stmt = db.select(Card).filter_by(id=id)
     card = (db.session.scalar(stmt))
     if card:
+        authorize(card.user_id)
         db.session.delete(card)
         db.session.commit()
         abort(200)
     else:
         return {'error':'Card not found'}, 404
+    
+cards_bp.register_blueprint(comments_bp)
